Fix ordinal encoding: emit floats so absent values stay missing, not rank 0
The M2 experiment surfaced an AutoGluon warning ("Int features with nulls:
imputing to 0"): integer ordinal codes made a fully-covered train column an
int-without-nulls, so AutoGluon imputed inference nulls to 0 — the WORST rank.
In House Prices, NA means "feature absent" (no pool/garage/basement) for many
quality columns, so this read "no pool" as "worst pool" and destroyed the
missingness signal. Emit float codes instead: NaN survives as missing and is
median-imputed. Regression test added; the experiment should be re-run for a
fair verdict.

+++ src/maestra/encoding.py
@@ -71,7 +71,11 @@ class OrdinalEncoding:
         out = df.copy()
         for col, mapping in self.maps.items():
             if col in out.columns:
-                out[col] = out[col].astype("object").map(mapping)
+                # float, never int: an unseen/absent value must stay NaN so AutoGluon treats it as
+                # missing (median-imputed), not as the lowest rank. As ints, a fully-covered train
+                # column becomes int-without-nulls and AutoGluon imputes inference nulls to 0 (the
+                # WORST level — e.g. "no pool" would read as "worst pool"), a systematic bias.
+                out[col] = out[col].astype("object").map(mapping).astype("float")
         return out
